changes to made everywhere it used on edit

+++ controller/customerController.js
@@ -5,6 +5,7 @@ const customerProductModel = require('../model/customerProductModel');
 const masterProductsModel = require('../model/masterProductsModel');
 const masterCustomerModel = require('../model/masterCustomerModel');
 const masterFirmModel = require('../model/masterFirmModel');
+const masterCategoriesModel = require('../model/masterCategoriesModel');
 
 
 
@@ -37,15 +38,35 @@ const postmasterFirm =(req,res,next)=>{
     const id = req.body.id;
     if(req.body.op==="del")
     {
-        masterFirmModel.destroy({
-            where: { firmId : id}
+        customerFirmModel.findAll(
+            {
+                where:{firm:id}
+            }
+        )
+        .then((data)=>{
+            console.log(data)
+            console.log(data.length)
+            if(data.length===0)
+            {
+                console.log("A")
+                masterFirmModel.destroy({
+                    where: { firmId : id}
+                })
+                .then((qwe)=>{
+                    res.send("kyu")
+                })
+                .catch((err)=>{
+                    console.log(err);
+                })  
+            }
+            else
+            {
+                console.log("B")
+                res.send("present")
+            }
         })
-        .then((qwe)=>{
-            res.send("kyu")
-        })
-        .catch((err)=>{
-            console.log(err);
-        })
+        .catch(err=>{console.log(err)})
+        
     }
     else if(req.body.op==="edt")
     {
@@ -155,10 +176,10 @@ const postCustomerContact =(req,res,next)=>{
     {
         console.log("new yr");
         customerContactModel.update({
-            customerName: req.body.customerName,
             designation: req.body.designation,
             mobileNumber: req.body.mobileNumber,
-            email: req.body.email
+            email: req.body.email,
+            ...(req.body.customerName !== "null" && { customerName: req.body.customerName })
           },
           {
               where: {customerId: req.body.id}
@@ -173,23 +194,41 @@ const postCustomerContact =(req,res,next)=>{
     }
     else
     {
-        const customerName = req.body.customerName;
-        const designation = req.body.designation;
-        const mobileNumber = req.body.mobileNumber;
-        const email = req.body.email;
-        customerContactModel.create({
-            customerName :customerName,
-            designation :designation,
-            mobileNumber :mobileNumber,
-            email :email
-        })
-        .then((result)=>{
-            console.log(result),
-            res.redirect('/customerContact')
-        })
-        .catch((err)=>
-            console.log(err)
-        )
+        const customerName = req.body.customerName
+        const contacts = req.body.contact || null
+        const contactNames = req.body.contactName
+        const designations = req.body.designation
+        const emails = req.body.email
+        
+        if(contacts)
+        {
+            console.log(contactNames)
+            console.log(contacts)
+            console.log(designations)
+            console.log(emails)
+            for (let i = 0; i < contacts.length; i++) {
+                const contactName = contactNames[i];
+                const contact = contacts[i];
+                const designation = designations[i];
+                const email = emails[i];
+            
+                customerContactModel
+                .create({
+                    customerName: customerId,
+                    mobileNumber: contact,
+                    contactName: contactName,
+                    designation: designation,
+                    email: email,
+                })
+                .then((res2) => {
+                    console.log(res2);
+                })
+                .catch((err) => {
+                    console.log(err);
+                });
+            }
+        }
+        res.redirect('/customerContact')
     }
     
 }
@@ -217,20 +256,27 @@ const getCustomerCategory =(req,res,next)=>{
         console.log(`User Level :  ${req.session.userLevel}`);
         masterCustomerModel.findAll()
         .then((custData)=>{
-
-            customerCategoryModel.findAll()
-            .then((data)=>{
-                console.log(data);
-                res.render('customerCategory',{
-                    username : req.session.username,
-                    data:data,
-                    custData: custData,
-                    level: req.session.userLevel
+            masterCategoriesModel.findAll()
+            .then((categorydata)=>{
+                customerCategoryModel.findAll()
+                .then((data)=>{
+                    console.log(data);
+                    res.render('customerCategory',{
+                        username : req.session.username,
+                        data:data,
+                        custData: custData,
+                        categorydata: categorydata,
+                        level: req.session.userLevel
+                    })
                 })
+                .catch((err)=>
+                console.log(err)
+                )
             })
             .catch((err)=>
             console.log(err)
             )
+            
         })
         .catch((err)=>
             console.log(err)
@@ -266,8 +312,9 @@ const postCustomerCategory =(req,res,next)=>{
     {
         console.log("new yr");
         customerCategoryModel.update({
-            customerName: req.body.customerName,
-            category: req.body.category
+            ...(req.body.customerName !== "null" && { customerName: req.body.customerName }),
+            ...(req.body.category !== "null" && { category: req.body.category })
+
           },
           {
               where: {customerId: req.body.id}
@@ -373,8 +420,9 @@ const postCustomerFirm =(req,res,next)=>{
     {
         console.log("new yr");
         customerFirmModel.update({
-            customerName: req.body.customerName,
-            firm: req.body.firm
+            ...(req.body.customerName !== "null" && { customerName: req.body.customerName }),
+            ...(req.body.firm !== "null" && { firm: req.body.firm })
+
           },
           {
               where: {customerId: req.body.id}
@@ -427,7 +475,6 @@ const getCustomerProduct =(req,res,next)=>{
         .then((custData)=>{
             masterProductsModel.findAll()
             .then((prodData)=>{
-                
                 customerProductModel.findAll()
                 .then((data)=>{
                     console.log(data);
@@ -483,8 +530,8 @@ const postCustomerProduct =(req,res,next)=>{
     {
         console.log("new yr");
         customerProductModel.update({
-            customerName: req.body.customerName,
-            product: req.body.product
+            ...(req.body.customerName !== "null" && { customerName: req.body.customerName }),
+            ...(req.body.product !== "null" && { product: req.body.product })
           },
           {
               where: {customerId: req.body.id}
